Наполнение базовых классов контроллера блоков Ver.3

+++ src/MyRoboMindMain/rMindBase/Interfaces/IDrawContainer.cs
@@ -2,7 +2,7 @@
 {
     public interface IDrawContainer : IDrawElement
     {
-        void SetPosition(float x, float y);
+        void SetPosition(double x, double y);
         void Translate(Types.Vector2 vector);
     }
 }
